Moving Config under io.zenoh
