+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/ConfigTest.kt
@@ -16,7 +16,6 @@ package io.zenoh
 import io.zenoh.exceptions.ZError
 import io.zenoh.keyexpr.intoKeyExpr
 import io.zenoh.bytes.into
-import io.zenoh.config.Config
 import io.zenoh.sample.Sample
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.delay
