+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/EncodingTest.kt
@@ -4,7 +4,6 @@ import io.zenoh.keyexpr.intoKeyExpr
 import io.zenoh.bytes.Encoding
 import io.zenoh.bytes.ZBytes
 import io.zenoh.bytes.into
-import io.zenoh.config.Config
 import io.zenoh.query.ReplyError
 import io.zenoh.sample.Sample
 import io.zenoh.query.intoSelector
