+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/GetTest.kt
@@ -17,7 +17,6 @@ package io.zenoh
 import io.zenoh.handlers.Handler
 import io.zenoh.sample.SampleKind
 import io.zenoh.bytes.into
-import io.zenoh.config.Config
 import io.zenoh.query.Reply
 import io.zenoh.query.Queryable
 import io.zenoh.query.Parameters
