+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/KeyExprTest.kt
@@ -14,7 +14,6 @@
 
 package io.zenoh
 
-import io.zenoh.config.Config
 import io.zenoh.keyexpr.KeyExpr
 import io.zenoh.keyexpr.SetIntersectionLevel
 import io.zenoh.keyexpr.intoKeyExpr
