+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/PublisherTest.kt
@@ -19,7 +19,6 @@ import io.zenoh.bytes.Encoding
 import io.zenoh.keyexpr.intoKeyExpr
 import io.zenoh.sample.SampleKind
 import io.zenoh.bytes.into
-import io.zenoh.config.Config
 import io.zenoh.pubsub.Publisher
 import io.zenoh.sample.Sample
 import io.zenoh.pubsub.Subscriber
