+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/PutTest.kt
@@ -17,7 +17,6 @@ package io.zenoh
 import io.zenoh.keyexpr.intoKeyExpr
 import io.zenoh.bytes.Encoding
 import io.zenoh.bytes.into
-import io.zenoh.config.Config
 import io.zenoh.sample.Sample
 import kotlin.test.Test
 import kotlin.test.assertEquals
