+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/QueryableTest.kt
@@ -19,7 +19,6 @@ import io.zenoh.handlers.Handler
 import io.zenoh.keyexpr.KeyExpr
 import io.zenoh.keyexpr.intoKeyExpr
 import io.zenoh.bytes.into
-import io.zenoh.config.Config
 import io.zenoh.qos.CongestionControl
 import io.zenoh.qos.Priority
 import io.zenoh.qos.QoS
