+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/SessionInfoTest.kt
@@ -14,7 +14,6 @@
 
 package io.zenoh
 
-import io.zenoh.config.Config
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
