+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/SubscriberTest.kt
@@ -23,7 +23,6 @@ import io.zenoh.qos.CongestionControl
 import io.zenoh.qos.Priority
 import io.zenoh.qos.QoS
 import io.zenoh.bytes.into
-import io.zenoh.config.Config
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.runBlocking
