+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/UserAttachmentTest.kt
@@ -19,7 +19,6 @@ import io.zenoh.keyexpr.intoKeyExpr
 import io.zenoh.bytes.ZBytes
 import io.zenoh.ext.zDeserialize
 import io.zenoh.bytes.into
-import io.zenoh.config.Config
 import io.zenoh.query.Reply
 import io.zenoh.sample.Sample
 import io.zenoh.query.Selector
